# Optimize thread pools by avoiding needless locks and unlocks of the `task_mutex`.