I profiled tps demo during gameplay and noticed 9.6% of time is spent on self time of `WorkerThreadPool::_thread_function`, specifically waiting for mutexes.

I found low hanging fruit to optimize: The current implementation needlessly unlocks and relocks the `task_mutex` on notification for the next task. This can be avoided by refactoring the function to have an inner loop, which avoids constructing `MutexLock lock` again each iteration, which is the cause of the needless unlock and relock.

## Profile Results

This optimization reduces the total CPU use of tps demo by about  2.3% (from `_thread_function`):
```
684.00 ms   9,6 %	0 s	  WorkerThreadPool::_thread_function(void*)
305.00 ms   4,3 %	0 s	  WorkerThreadPool::_process_task(WorkerThreadPool::Task*)
292.00 ms   4,1 %	0 s	  MDCommandBuffer::render_draw_indexed(unsigned int, unsigned int, unsigned int, int, unsigned int)
```
(current master)

```
736.00 ms   7,3 %	0 s	  WorkerThreadPool::_thread_function(void*)
477.00 ms   4,7 %	0 s	  MDCommandBuffer::render_draw_indexed(unsigned int, unsigned int, unsigned int, int, unsigned int)
453.00 ms   4,5 %	0 s	  WorkerThreadPool::_process_task(WorkerThreadPool::Task*)
```
(this PR)

My measurements didn't indicate a practical fps improvement. I think this is because this optimization does not affect the critical path. However, the total CPU load should still be lower than before.

## Addendum
I think `_wait_collaboratively` suffers from the same problem. It should probably be optimized in a future PR. 
However, its total CPU impact is much lower, and it is much more complex, so I didn't bother.